+++ pkg/team/serv.go
@@ -2,8 +2,8 @@ package team
 
 type Serv interface {
 	GetAll() ([]*Team, error)
-	GetByID(id uint64) (*Team, error)
-	Delete(id uint64) error
+	GetByID(id uint32) (*Team, error)
+	Delete(id uint32) error
 	Store(u *Team) error
 	Update(u *Team) error
 }
@@ -20,13 +20,13 @@ func NewTeamServ(repo Repo) Serv {
 
 func (svc *teamServ) DeleteByName(name string) error { return svc.repo.DeleteByName(name) }
 
-func (svc *teamServ) Delete(id uint64) error { return svc.repo.Delete(id) }
+func (svc *teamServ) Delete(id uint32) error { return svc.repo.Delete(id) }
 
 func (svc *teamServ) GetAll() ([]*Team, error) { return svc.repo.GetAll() }
 
 func (svc *teamServ) GetByName(name string) (*Team, error) { return svc.repo.GetByName(name) }
 
-func (svc *teamServ) GetByID(id uint64) (*Team, error) { return svc.repo.GetByID(id) }
+func (svc *teamServ) GetByID(id uint32) (*Team, error) { return svc.repo.GetByID(id) }
 
 func (svc *teamServ) Store(u *Team) error { return svc.repo.Store(u) }
